// src/awell/graphql-codegen/sdk.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  SafeDate: { input: any; output: any; }
};
// ...
export type DataPointInput = {
  data_point_definition_id: Scalars['String']['input'];
  value: Scalars['String']['input'];
};
// ...
export type StartPathwayInput = {
  data_points?: InputMaybe<Array<DataPointInput>>;
  pathway_definition_id: Scalars['String']['input'];
  patient_id: Scalars['String']['input'];
  release_id?: InputMaybe<Scalars['String']['input']>;
};